PROD4POD-658/Create matrix bubble chart
